Split unit tests into two files
